Add dropping items without closing INV with Right Click.
Adds the ability to right click instead of letting go of left click to drop item and that will allow the player to stay in their inventory to drop more items.
Also some disabled test code for quick slots inspired by Multiplayer but balanced for Ultimate Edition Reborn. Probably needs to be touched up to have the RNG factor of dropping it as well. Numbers will also display above the quick slots when an item is selectable.